requirements.txt
streamlit>=1.28.0
pandas==2.2.0
openpyxl==3.1.2
supabase>=2.8.0
httpx==0.27.2
python-dotenv==1.0.0
pydantic>=2.7.4
pygments==2.19.2
mdurl==0.1.2
markdown-it-py==4.0.0
rich>=10.14.0,<14
requests>=2.31.0
plotly>=5.17.0
openai>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
